Added node/text autoadjust.
Added the autoWidth, autoHeight and labelPadding Node configuration
parameters.
If autoWidth or autoHeight are true then the nodes dimensions will
adjust to the text inside them. This is calculated by preprocessing each
node and taking its "name" parameter and set the content of that
property to a hidden label that has "jit-autoadjust-label" as className.
Then the with and height is measured and setted as $width and $height
for that node.
The labelPadding param is useful for adding padding to the nodes so we
can be sure that the text remains inside the node.

This is a built-in solution to the problem described here:
http://groups.google.com/group/javascript-information-visualization-toolkit/browse_thread/thread/290cf12d062b055e/5947c184ce69fcb3