Added MouseWheel test and removed call to console
Now we're using event delegation for any type of events defined in
nodes/nodes labels. This means that, if before you created one
click/mouseover/out handler(s) for each label node and you had a graph
with 400 nodes, well, that would make something like c*400 listeners
(where c is a constant *not equal* to the speed of light). Event
delegation allows you to set just one listener, in the canvas element
itself, and when an event is fired we take advantage of the
capture/bubbling event system to capture the event in the canvas and
then grab the real target of the event and use that. So now if you have
a graph with 400 nodes c*400 is now just c.