Bug fix for calling indexOf method.
Line 142 checks the Array object definition for the indexOf method, and if found it calls the indexOf method on the passed array object.  There is nothing checking to make sure that array (the method parameter) is in fact an Array.  The parameter (array) should be checked for the indexOf method rather than the Array definition.