Fixed bugs related to morphing with directed graphs (specifically in RGraph)
The bugs appear when morphing from Graph A to Graph B, where A!=B, i.e.,
Where Edges in A either do not apear in B or in the reveres direction.

All bugs stem from the fact that rGraph is actually not-directed, and
the edge has a data memeber which holds the directions. Much of the
traversal operations used for calculations did not introspect this field
(either for deletion of edges nor for their reversal)
Example of bugs fixed:
* A comprises 2 nodes (x,y), and we morph from x->y to y->x., and back
(checking several morphs is a must since another bug was since must
operation traverse each edge twice, from each node).

Note that example only covers one of the bugs fixed in this commit.