Test Morph on Directed Graphs
Testing morph operations on a directed Graph, where edges are change
direction.
The edges' direction should be taken into considered when traversing the
Graph I.E., the BFS search should NOT follow a removed edge or a
reveresed edges (one that changes direction in the morph).
Running the tests: for each example, load the example clicking the
'Load' button followed by clicking the 'Run' button.