Updated offset calculation
Updated the offset calculation for the clicked node to ensure that the
canvas translation is considered.  For example with this change, when
Navigation.panning is true, and the graph is panned, the canvas
translation offsets will be considered when positioning the graph in
respect to the clicked node.  This change was derived from the following
post:
https://groups.google.com/d/topic/javascript-information-visualization-toolkit/1ElZw19VaXM/discussion.